Refactor render pipeline planning
